LayoutTests/canvas/philip/tests/2d.line.cap.closed.html

<!DOCTYPE html>
<title>Canvas test: 2d.line.cap.closed</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<link rel="prev" href="2d.line.cap.open.html" title="2d.line.cap.open">
<link rel="next" href="2d.line.cap.valid.html" title="2d.line.cap.valid">
<body class="show_output">
<p>
 <a href="2d.line.cap.open.html" accesskey="p" title="[p] 2d.line.cap.open">&lt;</a>
 <a href="index.html">[index]</a>
 <a href="2d.line.cap.valid.html" accesskey="n" title="[n] 2d.line.cap.valid">&gt;</a>
<h1><a href="index.2d.html">2d</a>.<a href="index.2d.line.html">line</a>.<a href="index.2d.line.cap.html">cap</a>.closed</h1>
<p class="desc">Line caps are not drawn at the corners of an unclosed rectangle</p>
<div class="refs">References:
<ul>
<li><a href="spec.html#testrefs.2d.lineCap.end">2d.lineCap.end</a>

</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#0f0';
ctx.strokeStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);

ctx.lineJoin = 'bevel';
ctx.lineCap = 'square';
ctx.lineWidth = 400;

ctx.beginPath();
ctx.moveTo(200, 200);
ctx.lineTo(200, 1000);
ctx.lineTo(1000, 1000);
ctx.lineTo(1000, 200);
ctx.closePath();
ctx.stroke();

_assertPixel(canvas, 1,1, 0,255,0,255, "1,1", "0,255,0,255");
_assertPixel(canvas, 48,1, 0,255,0,255, "48,1", "0,255,0,255");
_assertPixel(canvas, 48,48, 0,255,0,255, "48,48", "0,255,0,255");
_assertPixel(canvas, 1,48, 0,255,0,255, "1,48", "0,255,0,255");


});
</script>
