LayoutTests/canvas/philip/tests/2d.line.cap.closed-expected.txt

< [index] >
2d.line.cap.closed
Line caps are not drawn at the corners of an unclosed rectangle
References: 2d.lineCap.end
Actual output:
Expected output:

Failed assertion: got pixel [255,0,0,255] at (1,1), expected [0,255,0,255]
Failed assertion: got pixel [255,0,0,255] at (48,1), expected [0,255,0,255]
Failed assertion: got pixel [255,0,0,255] at (48,48), expected [0,255,0,255]
Failed assertion: got pixel [255,0,0,255] at (1,48), expected [0,255,0,255]
